+++ src/data/cakes.js
@@ -1,98 +1,26 @@
 const cakes = [
   {
-    title: 'Marble Cake',
+    title: 'Oxirate',
+    link: '/app/pages/prospect/5d2389f3cd1597865cb2cedc'
+  },
+  {
+    title: 'Invest in you',
     link: '/app/pages/details'
   },
   {
-    title: 'Fruitcake',
+    title: 'Capital Ventures',
     link: '/app/pages/details'
   },
   {
-    title: 'Chocolate Cake',
+    title: 'John Muir Hospital',
     link: '/app/pages/details'
   },
   {
-    title: 'Fat Rascal',
+    title: 'Health Financer',
     link: '/app/pages/details'
   },
   {
-    title: 'Financier',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Genoise',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Gingerbread',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Goose Breast',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Parkin',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Salzburger Nockerl',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Soufflé',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Merveilleux',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Streuselkuchen',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Tea Loaf',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Napoleonshat',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Merveilleux',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Magdalena',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Cremeschnitte',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Cheesecake',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Bebinca',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Fruitcake',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Chocolate Cake',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Fat Rascal',
-    link: '/app/pages/details'
-  },
-  {
-    title: 'Streuselkuchen',
+    title: 'Gives you money',
     link: '/app/pages/details'
   }
 ]
